requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.2
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
black==22.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
crispy-bootstrap5==0.6
dill==0.3.5.1
Django==4.1.10
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
docopt==0.6.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pipreqs==0.4.12
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pylint==2.14.1
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==7.1.2
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
yarg==0.1.9
